Extract EventForSelect as functional component
